Force fork-ts-checker to watch all TS files in the repo
There is an issue with fork-ts-checker (or perhaps we are misusing it) such that it does not re-typecheck the correct files when a change is made in the app, most notably for files that only export types. (https://github.com/TypeStrong/fork-ts-checker-webpack-plugin/issues/611)

This change forces it to look at these files in our package source directories rather than the symlinks in node_modules created by yarn workspaces. Now they are correctly watched and new type errors show up when a type-only file is edited.

fix build error

fix yarn.lock

Move base tsconfig changes to desktop tsconfigs